fix version set

+++ scripts/publish_lib.py
@@ -9,14 +9,16 @@ from build import ProjectBuilder
 from src.github_utils import create_release, upload_release_asset
 from src.version_utils import get_version
 
-def set_package_version(root_path: Path, version: int):
-    with open(root_path / 'pyproject.toml', 'r') as file:
+
+def set_package_version(version: int):
+    with open('pyproject.toml', 'r') as file:
         package_data = toml.load(file)
 
     package_data['project']['version'] = f'{version}.0.0'
 
-    with open(root_path / 'pyproject.toml', 'w') as file:
+    with open('pyproject.toml', 'w') as file:
         toml.dump(package_data, file)
+
 
 access_token = sys.argv[1]
 
@@ -26,14 +28,14 @@ if not access_token:
 
 (changed, version) = get_version(src=Path("src"), tag_prefix="version")
 
-# if not changed:
-#     exit()
+if not changed:
+    exit()
 
 rmtree("build", ignore_errors=True)
 rmtree("dist", ignore_errors=True)
 rmtree("publish_tools.egg-info", ignore_errors=True)
 
-set_package_version(Path("src"), version)
+set_package_version(version)
 
 ProjectBuilder(source_dir='.').build(
     distribution='wheel', output_directory='dist')
